# api/openapi.yaml
openapi: 3.1.0
info:
  title: Connector Service API
  description: The Connector Service provides a unified API for accessing Data Products,
    validating Contracts, and logging Transactions in the NextGen node.
  contact:
    name: HIRO-MicroDataCenters
    email: all-hiro@hiro-microdatacenters.nl
  license:
    name: MIT
    url: https://github.com/HIRO-MicroDataCenters-BV/ds-connector-service/blob/main/LICENSE
  version: 0.1.0
paths:
  /metrics:
    get:
      summary: Metrics
      description: Endpoint that serves Prometheus metrics.
      operationId: metrics_metrics_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /metadata/connector:
    get:
      tags:
      - Data Products
      summary: Get Connector Metadata
      description: 'partial(func, *args, **keywords) - new function with partial application

        of the given arguments and keywords.'
      operationId: get_connector_metadata
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorMetadata'
  /metadata/{interface_id}/{resource_path}/{resource_name}:
    get:
      tags:
      - Data Products
      summary: Get Data Product Distribution
      description: Return Metadata for a data product along with region.
      operationId: get_dataproduct_metadata
      parameters:
      - name: interface_id
        in: path
        required: true
        schema:
          type: string
          title: Interface Id
      - name: resource_path
        in: path
        required: true
        schema:
          type: string
          title: Resource Path
      - name: resource_name
        in: path
        required: true
        schema:
          type: string
          title: Resource Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /content/{interface_id}/{resource_path}/{resource_name}:
    get:
      tags:
      - Data Products
      summary: Get Data Product Content
      description: Return the full dataset content (CSV).
      operationId: get_dataproduct_content
      parameters:
      - name: interface_id
        in: path
        required: true
        schema:
          type: string
          title: Interface Id
      - name: resource_path
        in: path
        required: true
        schema:
          type: string
          title: Resource Path
      - name: resource_name
        in: path
        required: true
        schema:
          type: string
          title: Resource Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /content/{interface_id}/{resource_path}/{resource_name}/chunk:
    get:
      tags:
      - Data Products
      summary: Get Data Product Chunk
      description: Return a dataset chunk (partial CSV content).
      operationId: get_dataproduct_chunk
      parameters:
      - name: interface_id
        in: path
        required: true
        schema:
          type: string
          title: Interface Id
      - name: resource_path
        in: path
        required: true
        schema:
          type: string
          title: Resource Path
      - name: resource_name
        in: path
        required: true
        schema:
          type: string
          title: Resource Name
      - name: start
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Start byte position
          title: Start
        description: Start byte position
      - name: end
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: End byte position
          title: End
        description: End byte position
      responses:
        '200':
          description: Full chunk when no range specified
          content:
            application/json:
              schema: {}
        '206':
          description: Partial content chunk
          content:
            text/csv:
              example: 'partial,data

                '
          headers:
            Content-Range:
              description: Range of bytes returned
              schema:
                type: string
                example: bytes=0-1023
        '404':
          description: Data product not found
        '416':
          description: Range not satisfiable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /metadata-list/{interface_id}/{resource_path}:
    get:
      tags:
      - Data Products
      summary: List Data Products
      description: Return a paginated list of data product distributions with region.
      operationId: list_dataproducts
      parameters:
      - name: interface_id
        in: path
        required: true
        schema:
          type: string
          title: Interface Id
      - name: resource_path
        in: path
        required: true
        schema:
          type: string
          title: Resource Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataProductItem'
                title: Response List Dataproducts
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /health/{interface_id}:
    get:
      tags:
      - Data Products
      summary: Health Check
      description: 'partial(func, *args, **keywords) - new function with partial application

        of the given arguments and keywords.'
      operationId: health_check
      parameters:
      - name: interface_id
        in: path
        required: true
        schema:
          type: string
          title: Interface Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    ConnectorMetadata:
      properties:
        connector_id:
          type: string
          title: Connector Id
        region:
          type: string
          title: Region
        supported_interfaces:
          items:
            type: string
          type: array
          title: Supported Interfaces
        status:
          type: string
          title: Status
        version:
          type: string
          title: Version
      type: object
      required:
      - connector_id
      - region
      - supported_interfaces
      - status
      - version
      title: ConnectorMetadata
    DataProductDistribution:
      properties:
        access_service:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Service
        access_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Url
        byte_size:
          anyOf:
          - type: integer
          - type: 'null'
          title: Byte Size
        compress_format:
          anyOf:
          - type: string
          - type: 'null'
          title: Compress Format
        download_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Download Url
        media_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Media Type
        package_format:
          anyOf:
          - type: string
          - type: 'null'
          title: Package Format
        access_rights:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Rights
        conforms_to:
          anyOf:
          - type: string
          - type: 'null'
          title: Conforms To
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        format:
          anyOf:
          - type: string
          - type: 'null'
          title: Format
        issued:
          anyOf:
          - type: string
          - type: 'null'
          title: Issued
        license:
          anyOf:
          - type: string
          - type: 'null'
          title: License
        modified:
          anyOf:
          - type: string
          - type: 'null'
          title: Modified
        rights:
          anyOf:
          - type: string
          - type: 'null'
          title: Rights
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        has_policy:
          anyOf:
          - type: string
          - type: 'null'
          title: Has Policy
        checksum:
          anyOf:
          - type: string
          - type: 'null'
          title: Checksum
      type: object
      title: DataProductDistribution
      description: Standardized metadata for resources across different sources
    DataProductItem:
      properties:
        distribution:
          items:
            $ref: '#/components/schemas/DataProductDistribution'
          type: array
          title: Distribution
        region:
          type: string
          title: Region
          default: ''
      type: object
      required:
      - distribution
      title: DataProductItem
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError